🧪Test: (Refacor) WebSocket Scenario Test 설명 작성 및 클래스명 수정, inner 클래스 외부로 추출.
